Refactor conversation message UI
