+++ apps/studio/src/routes/editor/EditPanel/ChatTab/ChatMessages/AssistantMessage.tsx
@@ -1,0 +1,22 @@
+import type { AssistantChatMessageImpl } from '@/lib/editor/engine/chat/message/assistant';
+import { nanoid } from 'nanoid';
+import CodeChangeDisplay from '../CodeChangeDisplay';
+import MarkdownRenderer from './MarkdownRenderer';
+
+const AssistantMessage = ({ message }: { message: AssistantChatMessageImpl }) => {
+    return (
+        <div className="p-4 text-small content-start overflow-auto">
+            <div className="flex flex-col text-wrap gap-2">
+                {message.content.map((content) => {
+                    if (content.type === 'text') {
+                        return <MarkdownRenderer key={content.text} content={content.text} />;
+                    } else if (content.type === 'code') {
+                        return <CodeChangeDisplay key={nanoid()} content={content} />;
+                    }
+                })}
+            </div>
+        </div>
+    );
+};
+
+export default AssistantMessage;
